requirements.txt
black
fastapi
pandas
python-dotenv
requests
uvicorn